+++ package/src/functions/generate.ts
@@ -1,5 +1,6 @@
 import type * as Fs from "node:fs";
 
+import type { CopyTargetKind } from "#/@types/event";
 import type { Target, WriteFileData } from "#/@types/target";
 
 import * as Fsp from "node:fs/promises";
@@ -38,10 +39,8 @@ type GenerateTargetsOptions = {
     flatten: boolean;
 };
 
-type GeneratedTargetKind = "file" | "directory";
-
 type GeneratedTarget = {
-    kind: GeneratedTargetKind;
+    kind: CopyTargetKind;
     src: string;
     dest: string;
     renamed: boolean;
@@ -111,5 +110,5 @@ const generateTargets = async ({
     return result;
 };
 
-export type { GeneratedTarget, GeneratedTargetKind, GenerateTargetsOptions };
+export type { GeneratedTarget, GenerateTargetsOptions };
 export { generateTargets };