+++ package/src/index.ts
@@ -2,6 +2,11 @@ import { copy } from "#/plugins/copy";
 
 export default copy;
 
+export type {
+    CopyEvent,
+    CopyEventListener,
+    CopyTargetKind,
+} from "#/@types/event";
 export type { Options } from "#/@types/options";
 export type {
     Target,